requirements.txt
pandas
numpy
boto3
psycopg2-binary
apache-airflow
